refactor(services): remove AIService.test.ts and refactor AIService.ts
