+++ src/services/CacheManager.test.ts
@@ -1,0 +1,14 @@
+
+import { CacheManager } from "./CacheManager";
+
+describe("CacheManager", () => {
+	it("should generate a unique key", () => {
+		const cacheManager = new CacheManager(10);
+        const key = cacheManager.generateKey(
+            "test content", 
+            "test template",
+            "test style"
+        );
+        expect(key).toBe("test content:test template:test style");
+	});
+});
