+++ src/services/OpenAIService.test.ts
@@ -1,0 +1,45 @@
+import { OpenAIService } from './OpenAIService'; // Adjust the import path as necessary                                                                
+                                                                                                                                                        
+// Mock any dependencies here                                                                                                                          
+jest.mock('some-dependency', () => ({                                                                                                                  
+  // Mock implementation                                                                                                                               
+}));                                                                                                                                                   
+                                                                                                                                                       
+describe('OpenAIService', () => {                                                                                                                      
+  let openAIService: OpenAIService;                                                                                                                    
+                                                                                                                                                       
+  beforeEach(() => {                                                                                                                                   
+    openAIService = new OpenAIService('test-key', 'gpt-3.5-turbo');                                                                             
+  });                                                                                                                                                  
+                                                                                                                                                       
+  it('should analyze content successfully', async () => {                                                                                              
+    // Arrange: Set up any necessary conditions or inputs                                                                                              
+    const content = 'Test content';                                                                                                                    
+    const template = 'Test template';                                                                                                                  
+    const style = 'Test style';                                                                                                                        
+                                                                                                                                                       
+    // Mock the response from the service if needed                                                                                                    
+    jest.spyOn(openAIService, 'analyze').mockResolvedValue('Expected result');                                                                         
+                                                                                                                                                       
+    // Act: Call the method under test                                                                                                                 
+    const result = await openAIService.analyze(content, template, style);                                                                              
+                                                                                                                                                       
+    // Assert: Verify the result                                                                                                                       
+    expect(result).toBe('Expected result');                                                                                                            
+  });                                                                                                                                                  
+                                                                                                                                                       
+  it('should handle errors during analysis', async () => {                                                                                             
+    // Arrange                                                                                                                                         
+    const content = 'Test content';                                                                                                                    
+    const template = 'Test template';                                                                                                                  
+    const style = 'Test style';                                                                                                                        
+                                                                                                                                                       
+    // Mock an error response                                                                                                                          
+    jest.spyOn(openAIService, 'analyze').mockRejectedValue(new Error('Test error'));                                                                   
+                                                                                                                                                       
+    // Act & Assert: Verify that the method throws an error                                                                                            
+    await expect(openAIService.analyze(content, template, style)).rejects.toThrow('Test error');                                                       
+  });                                                                                                                                                  
+                                                                                                                                                       
+  // Add more test cases as needed                                                                                                                     
+});
